requirements.txt
sqlalchemy
psycopg2-binary
alembic
Faker
python-dotenv
